Fix extraction for dag based cost models
Report from Codex:

Recorded parent edges during relaxation so DAG-based costs keep their witness even when a child’s best cost later shrinks and would otherwise fail the final reconstruction.

Plan & Change

Strategies considered: 1) recompute parent edges with a fallback search when costs shift; 2) capture the winning hyperedge at the moment a cost improves to preserve a consistent witness. Chose option 2 to keep complexity low.

Now, when a cost update succeeds we store the corresponding hyperedge immediately (still requiring parent rank > child ranks) and removed the post-pass recomputation that could drop the witness for non-monotone costs (src/extract.rs).

+++ tests/integration_test.rs
@@ -1,6 +1,6 @@
 use egglog::{extract::DefaultCost, *};
 use egglog_ast::span::{RustSpan, Span};
-use std::collections::HashSet;
+use hashbrown::HashSet;
 
 #[test]
 fn globals_missing_prefix_warns_by_default() {
@@ -427,7 +427,7 @@ impl DagCostValue {
         let DagCostValue::Nodes(nodes) = self else {
             panic!("never will call")
         };
-        nodes.len() as usize
+        nodes.len()
     }
 }
